+++ src/GSS.Authorization.OAuth.HttpClient/OAuthHttpHandlerOptions.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Net.Http;
 using System.Threading.Tasks;
@@ -36,10 +35,5 @@ namespace GSS.Authorization.OAuth
         /// , see https://tools.ietf.org/html/rfc5849#section-3.5.2
         /// </summary>
         public bool SignedAsBody { get; set; }
-
-        /// <summary>
-        /// The form-encoded httpContent provider, see https://tools.ietf.org/html/rfc5849#section-3.5.2
-        /// </summary>
-        public Func<IEnumerable<KeyValuePair<string, string>>, HttpContent> FormUrlEncodedContentProvider { get; set; } = (values) => new FormUrlEncodedContent(values);
     }
 }