+++ src/GSS.Authorization.OAuth/OAuthOptions.cs
@@ -33,19 +33,6 @@ namespace GSS.Authorization.OAuth
         /// </summary>
         public string? Realm { get; set; }
 
-        [Obsolete("Use PercentEncoder instead")]
-        public Func<string?, string> PercentEncodeProvider
-        {
-            get
-            {
-                return (value) => value == null ? string.Empty : PercentEncoder(value);
-            }
-            set
-            {
-                PercentEncoder = value;
-            }
-        }
-
         /// <summary>
         /// The Percent-Encoder, see https://tools.ietf.org/html/rfc3986#section-2.1
         /// by default, the <see cref="Uri.EscapeDataString(string)"/> is RFC3986 compliant.