+++ test/GSS.Authorization.OAuth.Tests/FakeAuthorizer.cs
@@ -8,13 +8,15 @@ namespace GSS.Authorization.OAuth.Tests
 {
     internal class FakeAuthorizer : AuthorizerBase
     {
-        public FakeAuthorizer(IOptions<AuthorizerOptions> options, HttpClient httpClient, IRequestSigner signer) : base(options, httpClient, signer)
+        public FakeAuthorizer(IOptions<AuthorizerOptions> options, HttpClient httpClient, IRequestSigner signer) : base(
+            options, httpClient, signer)
         {
         }
 
         public string VerificationCode { get; set; }
 
-        public override Task<string> GetVerificationCodeAsync(Uri authorizeUri, CancellationToken cancellationToken = default)
+        public override Task<string> GetVerificationCodeAsync(Uri authorizeUri,
+            CancellationToken cancellationToken = default)
         {
             return Task.FromResult(VerificationCode);
         }