+++ test/GSS.Authorization.OAuth.Tests/RequestSignerTests.cs
@@ -1,7 +1,8 @@
 using System;
-using System.Collections.Specialized;
+using System.Collections.Generic;
 using System.Net.Http;
-using Microsoft.Extensions.Options;
+using Microsoft.AspNetCore.WebUtilities;
+using Microsoft.Extensions.Primitives;
 using Xunit;
 
 namespace GSS.Authorization.OAuth.Tests
@@ -15,24 +16,25 @@ namespace GSS.Authorization.OAuth.Tests
         public void GetBaseString()
         {
             // Arrange
-            var parameter = new NameValueCollection
+            var parameter = new Dictionary<string, StringValues>
             {
-                { "b5", "=%3D" },
-                { "a3", "a" },
-                { "c@", "" },
-                { "a2", "r b" },
-                { OAuthDefaults.OAuthConsumerKey, "9djdj82h48djs9d2" },
-                { OAuthDefaults.OAuthToken, "kkk9d7dh3k39sjv7" },
-                { OAuthDefaults.OAuthSignatureMethod, _signer.MethodName },
-                { OAuthDefaults.OAuthTimestamp, "137131201" },
-                { OAuthDefaults.OAuthNonce, "7d8f3e4a" },
-                { "c2", "" },
-                { "a3", "2 q" }
+                ["b5"] = "=%3D",
+                ["a3"] = new[] {"a", "2 q"},
+                ["c@"] = "",
+                ["a2"] = "r b",
+                [OAuthDefaults.OAuthConsumerKey] = "9djdj82h48djs9d2",
+                [OAuthDefaults.OAuthToken] = "kkk9d7dh3k39sjv7",
+                [OAuthDefaults.OAuthSignatureMethod] = _signer.MethodName,
+                [OAuthDefaults.OAuthTimestamp] = "137131201",
+                [OAuthDefaults.OAuthNonce] = "7d8f3e4a",
+                ["c2"] = ""
             };
-            var expected = "POST&http%3A%2F%2Fexample.com%2Frequest&a2%3Dr%2520b%26a3%3D2%2520q%26a3%3Da%26b5%3D%253D%25253D%26c%2540%3D%26c2%3D%26oauth_consumer_key%3D9djdj82h48djs9d2%26oauth_nonce%3D7d8f3e4a%26oauth_signature_method%3DHMAC-SHA1%26oauth_timestamp%3D137131201%26oauth_token%3Dkkk9d7dh3k39sjv7";
+            var expected =
+                "POST&http%3A%2F%2Fexample.com%2Frequest&a2%3Dr%2520b%26a3%3D2%2520q%26a3%3Da%26b5%3D%253D%25253D%26c%2540%3D%26c2%3D%26oauth_consumer_key%3D9djdj82h48djs9d2%26oauth_nonce%3D7d8f3e4a%26oauth_signature_method%3DHMAC-SHA1%26oauth_timestamp%3D137131201%26oauth_token%3Dkkk9d7dh3k39sjv7";
 
             // Act
-            var actual = ((RequestSignerBase)_signer).GetBaseString(HttpMethod.Post, new Uri("http://example.com/request?b5=%3D%253D&a3=a&c%40=&a2=r%20b"), parameter);
+            var actual = ((RequestSignerBase)_signer).GetBaseString(HttpMethod.Post,
+                new Uri("http://example.com/request?b5=%3D%253D&a3=a&c%40=&a2=r%20b"), parameter);
 
             // Assert
             Assert.Equal(expected, actual);
@@ -45,7 +47,7 @@ namespace GSS.Authorization.OAuth.Tests
             var expected = "74KNZJeDHnMBp0EMJ9ZHt/XKycU=";
             var clientCredentials = new OAuthCredential("dpf43f3p2l4k3l03", "kd94hf93k423kf44");
             var uri = new Uri("https://photos.example.net/initiate");
-            var parameters = new NameValueCollection
+            var parameters = new Dictionary<string, StringValues>
             {
                 [OAuthDefaults.Realm] = "Photos",
                 [OAuthDefaults.OAuthConsumerKey] = clientCredentials.Key,
@@ -72,7 +74,7 @@ namespace GSS.Authorization.OAuth.Tests
             var clientCredentials = new OAuthCredential("dpf43f3p2l4k3l03", "kd94hf93k423kf44");
             var temporaryCredentials = new OAuthCredential("hh5s93j4hdidpola", "hdhd0244k9j7ao03");
             var uri = new Uri("https://photos.example.net/token");
-            var parameters = new NameValueCollection
+            var parameters = new Dictionary<string, StringValues>
             {
                 [OAuthDefaults.Realm] = "Photos",
                 [OAuthDefaults.OAuthConsumerKey] = clientCredentials.Key,
@@ -101,7 +103,7 @@ namespace GSS.Authorization.OAuth.Tests
             var tokenCredentials = new OAuthCredential("nnch734d00sl2jdk", "pfkkdhi9sl3r4s00");
             var uri = new Uri("http://photos.example.net/photos?file=vacation.jpg&size=original");
 
-            var parameters = uri.ParseQueryString();
+            var parameters = QueryHelpers.ParseQuery(uri.Query);
             parameters[OAuthDefaults.OAuthConsumerKey] = clientCredentials.Key;
             parameters[OAuthDefaults.OAuthToken] = tokenCredentials.Key;
             parameters[OAuthDefaults.OAuthNonce] = "chapoH";